moved InvisibleItemFramesFeature.kt to misc feature
